+++ src/section4/Result.java
@@ -1,0 +1,7 @@
+package section4;
+
+enum Result {
+    WIN,
+    LOSE,
+    DRAW
+}
